requirements.txt
librosa==0.10.1
soundfile==0.12.1
Pillow==10.1.0
numpy==1.26.2
click==8.1.7
gradio==4.13.0
